Refactored basic auth mutations
