+++ backend/config/.env
@@ -1,0 +1,2 @@
+// IMPLEMENT DOTENV FILES!!!
+PRISMA_MANAGEMENT_API_SECRET=MggU5MdDRhD8UhIrQmRmgdKqclWLvqd3
