+++ backend/prisma/datamodel.prisma
@@ -3,7 +3,7 @@ type User {
   email: String! @unique
   passwordHash: String!
   name: String!
-  pesel: Int! @unique
+  pesel: String! @unique
   role: String!
   postalCode: String!
   votes: [Vote!]! @relation(name: "VoteToUser", onDelete: CASCADE)
