+++ backend/src/utils/customTypes.ts
@@ -4,3 +4,9 @@ export interface Context {
     request: ContextParameters;
     user?: { name: string; email: string; role: string };
 }
+
+export enum ROLE {
+    USER = 'user',
+    OPERATOR = 'operator',
+    ADMIN = 'admin'
+}
